// src/app/accounts/models/api/transaction.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionCategory } from './transactionCategory';
import { TransactionStatus } from './transactionStatus';

export interface Transaction { 
    transactionId?: string;
    name?: string;
    valueDate?: Date;
    bookedDate?: Date;
    amount?: number;
    currencyPair?: string;
    fxRate?: number;
    currencyCode?: string;
    isBooked?: boolean;
    statusName?: string;
    statusCode?: TransactionStatus;
    categoryCode?: TransactionCategory;
    categoryName?: string;
    otherPartyName?: string;
    otherPartyAccount?: string;
    otherPartyBank?: string;
    filboFee?: number;
    interchangeFee?: number;
    transactionReference?: string;
}