// src/app/accounts/models/api/transactionStatus.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Transaction status
 */
export type TransactionStatus = 'Pending' | 'Executed' | 'Declined';

export const TransactionStatus = {
    Pending: 'Pending' as TransactionStatus,
    Executed: 'Executed' as TransactionStatus,
    Declined: 'Declined' as TransactionStatus
};